Split graphics::Context into smaller sub-structs
I was getting anxiety around the manual clean up
in graphics::Context::drop so I just set things up
so that any given vulkan object has an Arc to its
relevant parent in a struct with it. That means if
we impl drop for the struct, we know that we're
destroying all children objects before we destroy
the parent. There's probably a better way to do
this that doesn't involve boxing everything with
an RC but I don't care right now and it'd probably
involved having more sophisticated systems (e.g. a
texture cache that manages your textures for you
probably won't need to pair every texture with a
pointer to the cache because those textures never
escape)